+++ CHANGELOG.md
@@ -19,4 +19,4 @@
 - [Improvements: Added node 6.10 support](https://github.com/agutoli/serverless-layers/commit/b13de031d8754591ee64b7e10b6d194156f02964)
 
 # 1.0.3 (2019-02-19)
-- [Bugfix: packagePath was not working](https://github.com/agutoli/serverless-layers/pull/1)
+- [Bugfix: dependenciesPath was not working](https://github.com/agutoli/serverless-layers/pull/1)